requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django==2.0.2
django-annoying==0.10.4
django-cors-headers==2.2.0
djangorestframework==3.7.7
Flask==0.12.2
Flask-Cors==3.0.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonpickle==0.9.6
MarkupSafe==1.0
numpy==1.14.1
opencv-python==3.4.0.12
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
